pyproject.toml
[project]
name = "itmo-chatbot"
version = "0.1.0"
description = "ITMO university programs scraper and RAG chatbot"
requires-python = ">=3.9"
dependencies = [
    "requests>=2.31.0",
    "beautifulsoup4>=4.12.0",
    "PyMuPDF>=1.23.0",
    "langchain>=0.3.0",
    "langchain-core>=0.3.0",
    "langchain-community>=0.3.0",
    "langchain-openai>=0.2.0",
    "faiss-cpu>=1.7.4",
    "sentence-transformers>=2.2.2",
    "openai>=1.0.0",
    "aiogram>=3.0.0",
    "scikit-learn>=1.3.0",
    "numpy>=1.24.0",
    "pandas>=2.0.0",
    "python-dotenv>=1.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]
include = [
    "*.py",
    "*.json",
    "*.md"
]
